+++ previous-versions/accounting/pkg/models/operations/downloadinvoicepdf.go
@@ -7,8 +7,9 @@ import (
 )
 
 type DownloadInvoicePdfRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
-	// Unique identifier for an invoice
+	// Unique identifier for an invoice.
 	InvoiceID string `pathParam:"style=simple,explode=false,name=invoiceId"`
 }
 
@@ -27,10 +28,13 @@ func (o *DownloadInvoicePdfRequest) GetInvoiceID() string {
 }
 
 type DownloadInvoicePdfResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Success
-	Data        []byte
-	StatusCode  int
+	Data []byte
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }