+++ previous-versions/accounting/pkg/models/operations/getcreditnote.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetCreditNoteRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a credit note.
 	CreditNoteID string `pathParam:"style=simple,explode=false,name=creditNoteId"`
 }
 
@@ -27,13 +29,16 @@ func (o *GetCreditNoteRequest) GetCreditNoteID() string {
 }
 
 type GetCreditNoteResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Success
 	CreditNote *shared.CreditNote
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetCreditNoteResponse) GetContentType() string {