+++ previous-versions/accounting/pkg/models/operations/refreshcompanyinfo.go
@@ -8,6 +8,7 @@ import (
 )
 
 type RefreshCompanyInfoRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 }
 
@@ -19,13 +20,16 @@ func (o *RefreshCompanyInfoRequest) GetCompanyID() string {
 }
 
 type RefreshCompanyInfoResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Success
 	Dataset *shared.Dataset
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *RefreshCompanyInfoResponse) GetContentType() string {