+++ previous-versions/accounting/RELEASES.md
@@ -371,3 +371,13 @@ Based on:
 - [go v0.26.0] previous-versions/accounting
 ### Releases
 - [Go v0.26.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.26.0 - previous-versions/accounting
+
+## 2023-09-26 00:20:59
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.91.0 (2.129.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.27.0] previous-versions/accounting
+### Releases
+- [Go v0.27.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.27.0 - previous-versions/accounting